Move WorkloadBudgetManager to core.
