+++ pinot-core/src/main/java/org/apache/pinot/core/accounting/PerQueryCPUMemAccountantFactory.java
@@ -56,8 +56,6 @@ import org.apache.pinot.spi.utils.CommonConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static org.apache.pinot.spi.utils.CommonConstants.Accounting.*;
-
 
 /**
  * Accounting mechanism for thread task execution status and thread resource usage sampling
@@ -335,7 +333,7 @@ public class PerQueryCPUMemAccountantFactory implements ThreadAccountantFactory 
     @Deprecated
     @Override
     public void setupRunner(String queryId, int taskId, ThreadExecutionContext.TaskType taskType) {
-      setupRunner(queryId, taskId, taskType, DEFAULT_WORKLOAD_NAME);
+      setupRunner(queryId, taskId, taskType, CommonConstants.Accounting.DEFAULT_WORKLOAD_NAME);
     }
